Made some UI Changes to better display the Invoice and incorporated DataTables to Incoice

+++ InventarVali.Models/Invoice.cs
@@ -15,8 +15,8 @@ namespace InventarVali.Models
         public int Id { get; set; }
         public string? InvoiceNumber { get; set; }
         public DateTime? InvoiceDate { get; set; }
-        public double? TotalPrice { get; set; }
-        public double Price { get; set; }
+        public decimal? TotalPrice { get; set; }
+        public decimal Price { get; set; }
         public string ?CardNumber { get; set; }
         public string? InvoiceUrl { get; set; }
         [ValidateNever]
